Fixes Android 12 text-only notification offset
Previously the text element was using a margin of the icon width, so that, even when hidden, the text element was offset by 64dp. Now using the relative layout to position the text element, and naming and hiding the frame when it is Android 12 style so that it is positioned correctly.