Allow server deep_link to result in an attributed launch
If client side deep link resolution fails, the launch_link is still sent to the server, and the server can resolve the link and return a deep_link. Previously this returned deep_link would get ignored if the client link resolution failed. It will now use the server's link-resolution and update the attribution properly.