Permit concurrent requests to users.json/return true from Session.setState when states are the same
The only place this is checked is when setting the state to IdentifyingUser before
submitting the users.json request. If the state is currently IdentifyingUser
(there is a users.json request in flight) then the request is dropped because the state
is already IdentifyingUser. This can result in player information updates not processing
if multiple happen within the same short timeframe, for example a push token being collected and
shortly after a player's email address being submitted (or a push token getting collected
shortly after the game provides Teak with initial information).