Fix tapping on notifications on Android <= 10
This fix existed in the 4.1.x series, but got partially reverted in 4.2+

The issue is that _something_ is using the PendingIntent requestCode as
a unique identifier for the bundle, so when there are two notifications
with the same PendingIntent requestCode, only the bundle from the most
recently received notification with that PendingIntent requestCode is used.

The 4.1.x series used a random number for all PendingIntents, but in the 4.2+
series this was reverted to only using a true random number for the Android 11+
case, while the Android <= 10 case using trampoline intents _could_ generate
duplicate requestCodes across multiple intents.

This applies the random number method to all cases, same as the 4.1.x series.
However, this should definitely be revisisted.